Move queue file
